Move, clean up, javadoc
