+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/graph/ui/LogFileWriter.java
@@ -1,0 +1,362 @@
+package org.nd4j.graph.ui;
+
+import com.google.flatbuffers.FlatBufferBuilder;
+import com.google.flatbuffers.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
+import org.nd4j.autodiff.samediff.SameDiff;
+import org.nd4j.base.Preconditions;
+import org.nd4j.graph.*;
+import org.nd4j.linalg.factory.Nd4j;
+import org.nd4j.linalg.primitives.Pair;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.RandomAccessFile;
+import java.nio.ByteBuffer;
+import java.nio.channels.FileChannel;
+import java.nio.channels.FileLock;
+import java.util.*;
+import java.util.concurrent.atomic.AtomicInteger;
+
+/**
+ * Log file writer - for writing append-only log file for UI etc.
+ *
+ * File format description: There are 2 blocks in the file
+ * 1. The "static information" block - containing 0 or more static info frames, followed by a UIStaticInfoRecord
+ *    with UIInfoType.START_EVENTS to signify the end of the "static information" section (start of events).
+ * 2. The "events" block - containing 0 or more event frames
+ *
+ * Note that once the UIInfoType.START_EVENTS UIStaticInfoRecord marker has been written, no more static information
+ * may be written; before this point, no event records may be written. This allows us to scan only the start of the
+ * file to get 'static' information such as the graph structure and hardware information.
+ *
+ *
+ * Data (for both sections) is recorded in "frames".<br>
+ * Each frame is formatted as follows:<br>
+ * [header_length_integer, content_length_integer, header_bytes, content_bytes]<br>
+ * These are as follows:
+ * <b>header_length_integer</b>: a signed 32-bit integer representing the number of bytes of the header data (stored in the "header_bytes" section of the frame)<br>
+ * <b>content_length_integer</b>: a signed 32-bit integer representing the number of bytes of the content data (stored in the "content_bytes" section of the frame).
+ *   May be 0 in some cases (such as when recording UIInfoType.START_EVENTS marker<br>
+ * <b>header_bytes</b>: flat-buffer encoded bytes representing either a UIStaticInfoRecord (in static information block) or a UIEvent (in the event block)
+ *   In both cases, this header object encodes the typo of data that follows.
+ * <b>content_bytes</b>: flat-buffer encoded bytes representing the content, of the type specified in the header<br>
+ * <br>
+ * <br>
+ *
+ * To-do list:<br>
+ * * Implement recording of remaining event types<br>
+ * * Handle loading already existing files<br>
+ * *
+ *
+ * @author Alex Black
+ */
+@Slf4j
+public class LogFileWriter {
+    private final File file;
+    private long endStaticInfoOffset = -1;
+    private final AtomicInteger nameIndexCounter = new AtomicInteger(0);
+    private final Map<Integer,String> nameIndexMap = new HashMap<>();
+    private final Map<String,Integer> indexNameMap = new HashMap<>();
+
+    public LogFileWriter(File file) throws IOException {
+        this.file = file;
+    }
+
+    /**
+     * Write the graph structure
+     * @param sd SameDiff instance to write the graph structure for
+     * @return Number of bytes written
+     * @throws IOException
+     */
+    public long writeGraphStructure(SameDiff sd) throws IOException {
+        Preconditions.checkState(endStaticInfoOffset < 0, "Cannot write graph structure - already wrote end of static info marker");
+        Pair<Integer, FlatBufferBuilder> h = encodeStaticHeader(UIInfoType.GRAPH_STRUCTURE);
+
+        FlatBufferBuilder fbb2 = new FlatBufferBuilder(0);
+        int graphStructureOffset = encodeGraphStructure(fbb2, sd);
+        fbb2.finish(graphStructureOffset);
+
+        return append(h.getSecond(), fbb2);
+    }
+
+
+    /**
+     * Write marker for final static data
+     * @return
+     * @throws IOException
+     */
+    public long writeFinishStaticMarker() throws IOException {
+        Preconditions.checkState(endStaticInfoOffset < 0, "Wrote final static already information already");
+        Pair<Integer, FlatBufferBuilder> encoded = encodeStaticHeader(UIInfoType.START_EVENTS);
+        long out = append(encoded.getSecond(), null);
+        endStaticInfoOffset = file.length();
+        return out;
+    }
+
+    /**
+     * Read all static information at the start of the file
+     *
+     * @return
+     * @throws IOException
+     */
+    public StaticInfo readStatic() throws IOException {
+
+        List<Pair<UIStaticInfoRecord, Table>> out = new ArrayList<>();
+        boolean allStaticRead = false;
+        try (RandomAccessFile f = new RandomAccessFile(file, "r"); FileChannel fc = f.getChannel()) {
+            f.seek(0);
+            while (!allStaticRead) {
+
+                //read 2 header ints
+                int lengthHeader = f.readInt();
+                int lengthContent = f.readInt();
+
+                //Read header
+                ByteBuffer bb = ByteBuffer.allocate(lengthHeader);
+                f.getChannel().read(bb);
+                bb.flip();      //Flip for reading
+                UIStaticInfoRecord r = UIStaticInfoRecord.getRootAsUIStaticInfoRecord(bb);
+
+                //Read content
+                bb = ByteBuffer.allocate(lengthContent);
+                f.getChannel().read(bb);
+                bb.flip();      //Flip for reading
+
+                byte infoType = r.infoType();
+                Table t;
+                switch (infoType) {
+                    case UIInfoType.GRAPH_STRUCTURE:
+                        t = UIGraphStructure.getRootAsUIGraphStructure(bb);
+                        break;
+                    case UIInfoType.SYTEM_INFO:
+                        t = UISystemInfo.getRootAsUISystemInfo(bb);
+                        break;
+                    case UIInfoType.START_EVENTS:
+                        t = null;
+                        break;
+                    default:
+                        throw new RuntimeException("Unknown UI static info type: " + r.infoType());
+                }
+
+                //TODO do we need to close file here?
+
+                out.add(new Pair<>(r, t));
+                long pointer = f.getFilePointer();
+                long length = f.length();
+                {
+                    log.trace("File pointer = {}, file length = {}", pointer, length);
+                    if (infoType == UIInfoType.START_EVENTS || pointer >= length) {
+                        allStaticRead = true;
+                    }
+                }
+            }
+            StaticInfo s = new StaticInfo(out, f.getFilePointer());
+            return s;
+        }
+    }
+
+    /**
+     * Read all of the events.
+     *
+     * @return All of the UI events
+     */
+    public List<Pair<UIEvent, Table>> readEvents() throws IOException {
+        //TODO eventually we'll support working out the offset for files that were not written in this session
+        Preconditions.checkState(endStaticInfoOffset >= 0, "Cannot read events - have not written end of static info marker");
+        return readEvents(endStaticInfoOffset);
+    }
+
+    /**
+     * Read all of the events starting at a specific file offset
+     *
+     * @return All of the UI events
+     */
+    public List<Pair<UIEvent, Table>> readEvents(long startOffset) throws IOException {
+        if(endStaticInfoOffset >= file.length()){
+            return Collections.emptyList();
+        }
+
+        List<Pair<UIEvent, Table>> out = new ArrayList<>();
+        try (RandomAccessFile f = new RandomAccessFile(file, "r"); FileChannel fc = f.getChannel()) {
+            f.seek(startOffset);
+            while (f.getFilePointer() < f.length()) {
+                //read 2 header ints
+                int lengthHeader = f.readInt();
+                int lengthContent = f.readInt();
+
+                //Read header
+                ByteBuffer bb = ByteBuffer.allocate(lengthHeader);
+                f.getChannel().read(bb);
+                bb.flip();      //Flip for reading
+                UIEvent e = UIEvent.getRootAsUIEvent(bb);
+
+                //Read Content
+                bb = ByteBuffer.allocate(lengthContent);
+                f.getChannel().read(bb);
+                bb.flip();      //Flip for reading
+
+                byte infoType = e.eventType();
+                Table t;
+                switch (infoType) {
+                    case UIEventType.ADD_NAME:
+                        t = UIAddName.getRootAsUIAddName(bb);
+                        break;
+                    case UIEventType.SCALAR:
+                    case UIEventType.ARRAY:
+                        t = FlatArray.getRootAsFlatArray(bb);
+                        break;
+
+                    //TODO
+                    case UIEventType.ARRAY_LIST:
+                    case UIEventType.HISTOGRAM:
+                    case UIEventType.IMAGE:
+                    case UIEventType.SUMMARY_STATISTICS:
+                    case UIEventType.OP_TIMING:
+                    case UIEventType.HARDWARE_STATE:
+                    case UIEventType.GC_EVENT:
+                    default:
+                        throw new RuntimeException("Unknown or not yet implemented event type: " + e.eventType());
+                }
+
+                //TODO do we need to close file here?
+
+                out.add(new Pair<>(e, t));
+            }
+            return out;
+        }
+    }
+
+    /**
+     * Register the event name - "accuracy", "loss", etc for later use in recording events.
+     * @param name Name to register
+     * @return Number of bytes written
+     */
+    public long registerEventName(String name) throws IOException {
+        Preconditions.checkState(endStaticInfoOffset >= 0, "Cannot write name - have not written end of static info marker");
+
+        FlatBufferBuilder fbb = new FlatBufferBuilder(0);
+        long time = System.currentTimeMillis();
+        int offset = UIEvent.createUIEvent(fbb, UIEventType.ADD_NAME, -1, time, 0, 0, (short)-1, 0, 0);
+        fbb.finish(offset);
+
+        FlatBufferBuilder fbb2 = new FlatBufferBuilder(0);
+        int idx = nameIndexCounter.getAndIncrement();
+        nameIndexMap.put(idx, name);
+        indexNameMap.put(name, idx);
+        int strOffset = fbb2.createString(name);
+        int offset2 = UIAddName.createUIAddName(fbb2, idx, strOffset);
+        fbb2.finish(offset2);
+
+        long l = append(fbb, fbb2);
+        return l;
+    }
+
+    /**
+     * Write a single scalar event to the file
+     * @param name      Name of the event. Must be registered by {@link #registerEventName(String)} first!
+     * @param time      Timestamp
+     * @param iteration Iteration for the event
+     * @param epoch     Epoch for the event
+     * @param scalar    Scalar value to write
+     * @return          Number of bytes written
+     */
+    public long writeScalarEvent(String name, long time, int iteration, int epoch, Number scalar) throws IOException {
+        //TODO add support for plugin and frame/iter
+        Preconditions.checkState(indexNameMap.containsKey(name), "Name \"%s\" not yet registered", name);
+        int idx = indexNameMap.get(name);
+        FlatBufferBuilder fbb = new FlatBufferBuilder(0);
+        int offset = UIEvent.createUIEvent(fbb, UIEventType.SCALAR, idx, time, iteration, epoch, (short)-1, 0, 0);
+        fbb.finish(offset);
+
+        FlatBufferBuilder fbb2 = new FlatBufferBuilder(0);
+        int offset2 = Nd4j.scalar(scalar).toFlatArray(fbb2);
+        fbb2.finish(offset2);
+
+        return append(fbb, fbb2);
+    }
+
+    /**
+     * Encode the header as a UIStaticInfoRecord instance for the specific {@link UIEventType}
+     * @param type UIEventType
+     */
+    protected Pair<Integer, FlatBufferBuilder> encodeStaticHeader(byte type) {
+        FlatBufferBuilder fbb = new FlatBufferBuilder(12);
+
+        int staticInfoOffset = UIStaticInfoRecord.createUIStaticInfoRecord(fbb, type);
+        fbb.finish(staticInfoOffset);
+        int lengthHeader = fbb.offset();       //MUST be called post finish to get real length
+        return new Pair<>(lengthHeader, fbb);
+    }
+
+    protected int encodeGraphStructure(FlatBufferBuilder fbb, SameDiff sd) {
+        //Create inputs list:
+        List<String> inputs = sd.inputs();
+        int[] inputListStrOffsets = new int[inputs.size()];
+        for (int i = 0; i < inputs.size(); i++) {
+            inputListStrOffsets[i] = fbb.createString(inputs.get(i));
+        }
+        int inputsOffset = UIGraphStructure.createInputsVector(fbb, inputListStrOffsets);
+
+        //Create inputs pair list:
+        int inputPairOffset = -1;
+
+        //Create outputs list:
+        List<String> outputs = sd.outputs();
+        int[] outputListStrOffsets = new int[outputs.size()];
+        for (int i = 0; i < outputListStrOffsets.length; i++) {
+            outputListStrOffsets[i] = fbb.createString(outputs.get(i));
+        }
+        int outputsOffset = UIGraphStructure.createInputsVector(fbb, outputListStrOffsets);
+
+        //Create variables list
+        int outputsListOffset = -1;
+
+        //Create ops list
+        int opsListOffset = -1;
+
+        return UIGraphStructure.createUIGraphStructure(fbb, inputsOffset, inputPairOffset, outputsOffset, outputsListOffset, opsListOffset);
+    }
+
+    private long append(FlatBufferBuilder h, FlatBufferBuilder c) throws IOException {
+        ByteBuffer bb1 = h.dataBuffer();
+        ByteBuffer bb2 = (c == null ? null : c.dataBuffer());
+
+        try (RandomAccessFile f = new RandomAccessFile(file, "rw"); FileChannel fc = f.getChannel(); FileLock lock = fc.lock()) {
+            //TODO can we make this more efficient - use a single byte buffer?
+
+            //Seek to end for append
+            f.seek(f.length());
+            long startPos = f.getFilePointer();
+
+            //Write header - length of SystemInfo header, length of content header
+            ByteBuffer header = ByteBuffer.allocate(8); //8 bytes = 2x 4 byte integers
+            int l1 = bb1.remaining();
+            int l2 = bb2 == null ? 0 : bb2.remaining();
+            header.putInt(l1);
+            header.putInt(l2);
+            header.flip();
+
+            //System.out.println("Lengths - header, content: " + l1 + ", " + l2);
+
+            int w1 = fc.write(header);
+            int w2 = fc.write(bb1);
+            int w3 = bb2 == null ? 0 : fc.write(bb2);
+            long total = w1 + w2 + w3;
+            //System.out.println("Wrote " + total + " bytes starting at position " + startPos);
+            //System.out.println("Post writing file length: " + file.length());
+            return total;
+        }
+    }
+
+
+
+    @AllArgsConstructor
+    @Data
+    public static class StaticInfo {
+        private final List<Pair<UIStaticInfoRecord, Table>> data;
+        private final long fileOffset;
+    }
+}
